重构一波，为了让代码更简单
