+++ RenderPass.h
@@ -5,8 +5,10 @@
 #ifndef RENDERPASS_H
 #define RENDERPASS_H
 #include <vulkan/vulkan.h>
+#include "LLVK_SYS.hpp"
 
 
+LLVK_NAMESPACE_BEGIN
 struct RenderPass {
     VkDevice bindDevice{}; // bind device
     VkPhysicalDevice bindPhysicalDevice{};
@@ -16,7 +18,7 @@ struct RenderPass {
     void init();
     void cleanup();
 };
-
+LLVK_NAMESPACE_END
 
 
 #endif //RENDERPASS_H
